flaskTest/requirements.txt
mysql-connector-python==8.1.0
click==8.1.7
flask==2.3.3
werkzeug==2.3.7